requirements.txt
cryptography>=41.0.0
pycryptodome>=3.19.0
pydantic>=2.5.0
pyyaml>=6.0.1
fastapi>=0.109.0
uvicorn>=0.27.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
aiofiles>=23.2.1
numpy>=1.26.0
